search-engine/requirements.txt
curl_cffi==0.13.0
zeroapi==0.9.0
python-dotenv==1.1.1
tenacity==9.1.2
limits==5.6.0
tldextract==5.3.0
lxml==6.0.2
cssselect==1.3.0
trafilatura==2.0.0
sentence-transformers==5.1.1
hnswlib==0.8.0
pytest==8.4.2
pytest-asyncio==1.2.0